Add .gitignore
* add .gitignore
* remove superfluous obj directories - these will be rebuilt by VS.NET
at each build.

Yes, I know the .gitignore is ridiculously big, and covers a huge swatch
of likely unneccessary stanzas. I figured it would be best to start with
a baseline that included the kitchen sink - and then pair things down
once it's clear what people will and will not use to help contribute to
this repo.